Предварительный вариант, на проверку.
