+++ composer.json
@@ -1,0 +1,18 @@
+{
+  "name": "alex/task-force",
+  "description": "Учебный проект htmlacademy",
+  "type": "project",
+  "authors": [
+    {
+      "name": "Alexey Pozhidaev",
+      "email": "worka@bk.ru"
+    }
+  ],
+  "minimum-stability": "dev",
+  "require": {},
+  "autoload": {
+    "psr-4": {
+      "app\\": "src/"
+    }
+  }
+}
