+++ src/exceptions/base/TaskActionException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace app\exceptions\base;
+
+/**
+ * Базовое исключение "Состояний" в "Заданиях"
+ */
+class TaskActionException extends TaskException
+{
+}
