+++ src/exceptions/base/TaskException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace app\exceptions\base;
+
+class TaskException extends \Exception
+{
+}
