+++ src/exceptions/base/TaskRoleException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace app\exceptions\base;
+
+/**
+ * Базовое исключение "Ролей" в "Заданиях"
+ */
+class TaskRoleException extends TaskException
+{
+}
