+++ src/exceptions/base/TaskStatusException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace app\exceptions\base;
+
+/**
+ * Базовое исключение "Состояний" в "Заданиях"
+ */
+class TaskStatusException extends TaskException
+{
+}
