+++ src/exceptions/task/NotValidStatusException.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace app\exceptions\task;
+
+use app\exceptions\base\TaskStatusException;
+
+class NotValidStatusException extends TaskStatusException
+{
+    /**
+     * @param string $status "Состояние"
+     * @param int $code Код исключения, по умолчанию 0
+     */
+    public function __construct(string $status, int $code = 0)
+    {
+        $message = "Не допустимое значение \"Состояния\" - {$status}";
+        parent::__construct($message, $code);
+    }
+}
