+++ src/models/User.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace app\models;
+
+class User
+{
+}
